部署设计页面的删除功能添加确认功能

+++ src/pages/deployplan.vue
@@ -185,52 +185,60 @@ export default{
 
                 var id = comptrInfo.id;
 
+                var msg = "您确定删除吗？";
+                if (confirm(msg) == true) {
 
-                if (target.parentNode.parentNode.parentNode.tagName.toLowerCase() == "td") {
-                    //alert("C");
-                    var rowIndex = target.parentNode.parentNode.parentNode.parentNode.rowIndex;
-                    //alert(rowIndex);
-                    var id = document.getElementById("table_value").rows[rowIndex].cells[0].innerHTML;
-                    //alert(id);
-                    var qs = require('qs');
-                    this.$axios.delete('deployplan/'+id,{
+                    if (target.parentNode.parentNode.parentNode.tagName.toLowerCase() == "td") {
+                        //alert("C");
+                        var rowIndex = target.parentNode.parentNode.parentNode.parentNode.rowIndex;
+                        //alert(rowIndex);
+                        var id = document.getElementById("table_value").rows[rowIndex].cells[0].innerHTML;
+                        //alert(id);
+                        var qs = require('qs');
+                        this.$axios.delete('deployplan/'+id,{
 
-                        //设置头
-                        headers:{
-                            'content-type':'application/x-www-form-urlencoded'
-                        },
-                        auth: {
-                            username: 'admin',
-                            password: 'admin'
-                        }
-                    }).then(res=>{
-                        
-                        //删除完再次查询
-                        var projectId = this.getCookie('projectId');
-                        var username = this.getCookie('username');
-                        var password = this.getCookie('password');
-                        
-                        this.$axios.get('project/'+projectId+'/deployplan',{
                             //设置头
                             headers:{
                                 'content-type':'application/x-www-form-urlencoded'
                             },
                             auth: {
-                                username: username,
-                                password: password
+                                username: 'admin',
+                                password: 'admin'
                             }
                         }).then(res=>{
-                            this.deployplans = res.data.data;
-                            console.log(this.deployplans);
-                        })
-                        .catch(err=>{
-                            console.log(err);
-                        })
+                            
+                            //删除完再次查询
+                            var projectId = this.getCookie('projectId');
+                            var username = this.getCookie('username');
+                            var password = this.getCookie('password');
+                            
+                            this.$axios.get('project/'+projectId+'/deployplan',{
+                                //设置头
+                                headers:{
+                                    'content-type':'application/x-www-form-urlencoded'
+                                },
+                                auth: {
+                                    username: username,
+                                    password: password
+                                }
+                            }).then(res=>{
+                                this.deployplans = res.data.data;
+                                console.log(this.deployplans);
+                            })
+                            .catch(err=>{
+                                console.log(err);
+                            })
 
-                    }).catch(err=>{
-                        alert("删除失败！");
-                    })
+                        }).catch(err=>{
+                            alert("删除失败！");
+                        })
+                    }
+                    
+                } else {
+                    return false;
                 }
+
+                
 
             },
     }
